[Refactor]: dto/groupJoin 패키지로 이전
